add model API landmarks to gaze_estimation_demo

+++ demos/gaze_estimation_demo/cpp/include/eye_state_estimator.hpp
@@ -24,7 +24,7 @@ public:
     const std::string modelType = "Eye State Estimation";
 
 private:
-    cv::Rect createEyeBoundingBox(const cv::Point2i& p1, const cv::Point2i& p2, float scale = 1.8) const;
+    cv::Rect createEyeBoundingBox(const cv::Point2f& p1, const cv::Point2f& p2, float scale = 1.8) const;
     void rotateImageAroundCenter(const cv::Mat& srcImage, cv::Mat& dstImage, float angle) const;
 
     IEWrapper ieWrapper;
